+++ next.config.ts
@@ -3,7 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   /* config options here */
   images: {
-    domains: ["xdijpvmnapufbkbsmhwn.supabase.co"],
+    domains: ["xdijpvmnapufbkbsmhwn.supabase.co", "lh3.googleusercontent.com"],
   },
 };
